# apps/server/public/openApi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/hardware/badge/callback": {
      "post": {
        "operationId": "SmartBadgeController_callback",
        "summary": "工牌回调通知",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BadgeCallbackDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "硬件-智能工牌"
        ]
      }
    },
    "/hardware/badge/bind": {
      "post": {
        "operationId": "SmartBadgeController_bindDevice",
        "summary": "绑定设备到用户",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BindDeviceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "硬件-智能工牌"
        ]
      }
    },
    "/hardware/badge/device": {
      "post": {
        "operationId": "SmartBadgeController_createDevice",
        "summary": "新增设备主档",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBadgeDeviceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "硬件-智能工牌"
        ]
      }
    },
    "/hardware/badge/device/update": {
      "post": {
        "operationId": "SmartBadgeController_updateDevice",
        "summary": "编辑设备主档",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBadgeDeviceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "硬件-智能工牌"
        ]
      }
    },
    "/hardware/badge/device/delete-review": {
      "post": {
        "operationId": "SmartBadgeController_deleteDevice",
        "summary": "多级校验删除设备主档",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteBadgeDeviceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "硬件-智能工牌"
        ]
      }
    },
    "/hardware/badge/device/restore/{id}": {
      "post": {
        "operationId": "SmartBadgeController_restoreDevice",
        "summary": "恢复已删除设备主档",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "硬件-智能工牌"
        ]
      }
    },
    "/hardware/badge/unbind": {
      "post": {
        "operationId": "SmartBadgeController_unbindDevice",
        "summary": "解绑设备",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnbindDeviceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "硬件-智能工牌"
        ]
      }
    },
...
    "/hardware/badge/record/{id}": {
      "delete": {
        "operationId": "SmartBadgeController_recordDelete",
        "summary": "删除录音记录",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "硬件-智能工牌"
        ]
      }
    },
    "/hardware/badge/transcribe/{id}": {
      "post": {
        "operationId": "SmartBadgeController_manualTranscribe",
        "summary": "手动触发阿里云 ASR 转写",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "硬件-智能工牌"
        ]
      }
    },
...
    "/hardware/badge/audit/report/detail/{id}": {
      "get": {
        "operationId": "SmartBadgeController_reportDetail",
        "summary": "AI 日报详情，包含大文本",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "硬件-智能工牌"
        ]
      }
    },
...
    "/tenant-care/badge/bind": {
      "post": {
        "operationId": "TenantCareController_bindBadge",
        "summary": "租户护工绑定工牌",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BindTenantBadgeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "租户护理"
        ]
      }
    },
    "/tenant-care/badge/unbind": {
      "post": {
        "operationId": "TenantCareController_unbindBadge",
        "summary": "租户护工解绑工牌",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnbindTenantBadgeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "租户护理"
        ]
      }
    },
    "/tenant-care/badge/binding/list": {
      "get": {
        "operationId": "TenantCareController_listBindings",
        "summary": "租户工牌绑定列表",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "description": "当前分页",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "orderByColumn",
            "required": false,
            "in": "query",
            "description": "排序字段",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isAsc",
            "required": false,
            "in": "query",
            "description": "排序规则",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantId",
            "required": false,
            "in": "query",
            "description": "平台管理员指定租户ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deviceNo",
            "required": false,
            "in": "query",
            "description": "设备编号",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "护工姓名/手机号",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isCurrent",
            "required": false,
            "in": "query",
            "description": "是否当前有效绑定",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "租户护理"
        ]
      }
    },
...
  },
  "info": {
    "title": "Nest-Admin",
    "description": "Nest-Admin 接口文档",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:8081"
    }
  ],
  "components": {
    "securitySchemes": {
      "Authorization": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "description": "请在请求头中携带 JWT 令牌，格式：Bearer <token>"
      }
    },
    "schemas": {
...
      "BadgeCallbackDto": {
        "type": "object",
        "properties": {
          "dataType": {
            "type": "string",
            "description": "数据类型: Audio|Gps|AudioText|LoginLog|ControlLog|UploadLog|HeartbeatLog|DebugLog|MergeAudio"
          },
          "timeStamp": {
            "type": "string",
            "description": "时间戳"
          },
          "deviceNo": {
            "type": "string",
            "description": "设备编号"
          },
          "data": {
            "type": "object",
            "description": "业务数据体"
          }
        },
        "required": [
          "dataType",
          "timeStamp",
          "deviceNo",
          "data"
        ]
      },
      "BindDeviceDto": {
        "type": "object",
        "properties": {
          "deviceNo": {
            "type": "string",
            "description": "设备号"
          },
          "userId": {
            "type": "number",
            "description": "用户 ID"
          }
        },
        "required": [
          "deviceNo",
          "userId"
        ]
      },
      "CreateBadgeDeviceDto": {
        "type": "object",
        "properties": {
          "deviceNo": {
            "type": "string",
            "description": "设备号"
          },
          "status": {
            "type": "string",
            "description": "状态 0启用 1停用"
          },
          "remark": {
            "type": "string",
            "description": "备注"
          }
        },
        "required": [
          "deviceNo"
        ]
      },
      "UpdateBadgeDeviceDto": {
        "type": "object",
        "properties": {
          "deviceNo": {
            "type": "string",
            "description": "设备号"
          },
          "status": {
            "type": "string",
            "description": "状态 0启用 1停用"
          },
          "remark": {
            "type": "string",
            "description": "备注"
          },
          "id": {
            "type": "number",
            "description": "设备主键"
          }
        },
        "required": [
          "deviceNo",
          "id"
        ]
      },
      "DeleteBadgeDeviceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "设备主键"
          },
          "confirmDeviceNo": {
            "type": "string",
            "description": "二次确认的设备号"
          },
          "reason": {
            "type": "string",
            "description": "删除原因"
          }
        },
        "required": [
          "id",
          "confirmDeviceNo",
          "reason"
        ]
      },
      "UnbindDeviceDto": {
        "type": "object",
        "properties": {
          "deviceNo": {
            "type": "string",
            "description": "设备号"
          }
        },
        "required": [
          "deviceNo"
        ]
      },
...
      "BindTenantBadgeDto": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "description": "平台管理员指定租户ID"
          },
          "tenantCaregiverId": {
            "type": "number",
            "description": "租户护工ID"
          },
          "deviceNo": {
            "type": "string",
            "description": "设备编号"
          }
        },
        "required": [
          "tenantCaregiverId",
          "deviceNo"
        ]
      },
      "UnbindTenantBadgeDto": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "description": "平台管理员指定租户ID"
          },
          "deviceNo": {
            "type": "string",
            "description": "设备编号"
          }
        },
        "required": [
          "deviceNo"
        ]
      },
...
    }
  }
}